Change distance type to double wherever there was long

+++ app/src/main/java/com/actively/activity/Activity.kt
@@ -49,7 +49,7 @@ data class Location(
         this.toPoint(),
         other.toPoint(),
         TurfConstants.UNIT_METERS
-    ).toLong().meters
+    ).meters
 }
 
 fun AndroidLocation.toLocation() = Location(
